moved slice functions to op_slice
